+++ packages/perseus/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @khanacademy/perseus
 
+## 82.0.1
+
+### Patch Changes
+
+-   [#3836](https://github.com/Khan/perseus/pull/3836) [`3ade343`](https://github.com/Khan/perseus/commit/3ade34395ead8a25f2dc1e909275d736b6b4298a) Thanks [@Myranae](https://github.com/Myranae)! - Convert the `graded-group`, `graded-group-answer-bar`, and `graded-group-set` widgets to use Wonder Blocks semantic color and typography tokens. These widgets now follow the active theme (including dark mode).
+
+-   [#3898](https://github.com/Khan/perseus/pull/3898) [`a23cf05`](https://github.com/Khan/perseus/commit/a23cf05c82bc9ef3e78f8f11b4b8b44d8a99ca37) Thanks [@ivyolamit](https://github.com/ivyolamit)! - Add asympote to tangent interactive graph
+
+-   [#3946](https://github.com/Khan/perseus/pull/3946) [`839e471`](https://github.com/Khan/perseus/commit/839e471be4108f6103cede4d1e7f9391872de519) Thanks [@benchristel](https://github.com/benchristel)! - Internal: the Number Line widget has been refactored.
+
+-   [#3951](https://github.com/Khan/perseus/pull/3951) [`38d9791`](https://github.com/Khan/perseus/commit/38d9791debec5865a3e4eaf5801323636a8516d5) Thanks [@ivyolamit](https://github.com/ivyolamit)! - [Interactive Graph] Fix editor crash when point graph is set to unlimited (0/unlimited points), and deleted unused '0 points' dropdown selection
+
+## 82.0.0
+
+### Major Changes
+
+-   [#3936](https://github.com/Khan/perseus/pull/3936) [`2a0da4f`](https://github.com/Khan/perseus/commit/2a0da4f190dce68d12b04eef13f84698b18b4ebd) Thanks [@benchristel](https://github.com/benchristel)! - Make all options of the Plotter widget required, and default them in the parser.
+
+### Minor Changes
+
+-   [#3937](https://github.com/Khan/perseus/pull/3937) [`469b559`](https://github.com/Khan/perseus/commit/469b55948daa2897756a83842378ff74cadb5037) Thanks [@nishasy](https://github.com/nishasy)! - [Matrix] Stay left-to-right, even in right-to-left languages
+
+-   [#3932](https://github.com/Khan/perseus/pull/3932) [`853d839`](https://github.com/Khan/perseus/commit/853d83947bb84692610d7691c9f8b1156f5187e0) Thanks [@benchristel](https://github.com/benchristel)! - The Molecule widget (already deprecated and hidden from the editor) has been removed. It now renders as the DeprecatedStandin widget.
+
+-   [#3943](https://github.com/Khan/perseus/pull/3943) [`0c19685`](https://github.com/Khan/perseus/commit/0c196858400547eb3180747fc4bfc34aee9dd684) Thanks [@EmiliaPalaghita](https://github.com/EmiliaPalaghita)! - Remove the `perseus-enable-point-label-field` feature flag (AITQ-385). `showPointLabels` is now honored unconditionally on interactive graphs and in the editor toggle.
+
+-   [#3890](https://github.com/Khan/perseus/pull/3890) [`4278b68`](https://github.com/Khan/perseus/commit/4278b68ab445c90c8efeb23924ffce5d9471dc73) Thanks [@catandthemachines](https://github.com/catandthemachines)! - [Interactive Graph] Update strings for more dynamic point labels.
+
+### Patch Changes
+
+-   [#3938](https://github.com/Khan/perseus/pull/3938) [`bfb1191`](https://github.com/Khan/perseus/commit/bfb11914e4052fe1938e55f75732b742533f37d1) Thanks [@benchristel](https://github.com/benchristel)! - Internal: The PythonProgram widget no longer assigns defaults for props, since its options are never undefined in the data.
+
+-   [#3933](https://github.com/Khan/perseus/pull/3933) [`0e0abbc`](https://github.com/Khan/perseus/commit/0e0abbc84f4c65a3b1750102f76823a7746f97c0) Thanks [@benchristel](https://github.com/benchristel)! - Internal: the unused default props of the Measurer widget have been removed.
+
+-   [#3939](https://github.com/Khan/perseus/pull/3939) [`53d364e`](https://github.com/Khan/perseus/commit/53d364e51313887960ebec480a10d97f70a54d57) Thanks [@benchristel](https://github.com/benchristel)! - Internal: `defaultProps` have been removed from the `FreeResponse` widget. The `userInput` prop did not need a default; it is always passed.
+
+-   [#3934](https://github.com/Khan/perseus/pull/3934) [`da5054f`](https://github.com/Khan/perseus/commit/da5054fb310803a9db446052533121dbce3a149f) Thanks [@benchristel](https://github.com/benchristel)! - Internal: Remove unused default props from the NumberLine widget. All required options are defaulted by the parser.
+
+-   Updated dependencies [[`2a0da4f`](https://github.com/Khan/perseus/commit/2a0da4f190dce68d12b04eef13f84698b18b4ebd), [`853d839`](https://github.com/Khan/perseus/commit/853d83947bb84692610d7691c9f8b1156f5187e0), [`0e0abbc`](https://github.com/Khan/perseus/commit/0e0abbc84f4c65a3b1750102f76823a7746f97c0), [`0c19685`](https://github.com/Khan/perseus/commit/0c196858400547eb3180747fc4bfc34aee9dd684), [`e50990e`](https://github.com/Khan/perseus/commit/e50990e172e0c6a28c69384a3a0f01240694efc6)]:
+    -   @khanacademy/perseus-core@33.0.0
+    -   @khanacademy/perseus-linter@5.1.16
+    -   @khanacademy/keypad-context@3.2.76
+    -   @khanacademy/kmath@2.4.35
+    -   @khanacademy/math-input@26.6.5
+    -   @khanacademy/perseus-score@8.12.8
+
+## 81.0.2
+
+### Patch Changes
+
+-   Updated dependencies [[`6db4fc8`](https://github.com/Khan/perseus/commit/6db4fc87370fdd2488a14a9eab514a22b6ad3dcc)]:
+    -   @khanacademy/math-input@26.6.4
+
 ## 81.0.1
 
 ### Patch Changes